[imporve]: plit discovery center related monitoring into a separate module
